refactor: close transaction purity standards findings
