+++ packages/shared/services/QueueBoxUtilities.ts
@@ -1,6 +1,5 @@
 import { Temporal } from '@js-temporal/polyfill';
 import { ALMessage } from '../al-contracts/al-contract.ts';
-import { computeResourceEntryFromALMessage } from '../queuebox/al-message-resource-entry-computation.ts';
 import { DequeueController } from '../queuebox/DequeueController.ts';
 import {
     DequeueResourceEntryController,
@@ -9,6 +8,7 @@ import {
 } from '../queuebox/DequeueResourceEntryController.ts';
 import { QueueBoxResourceEntryRepository } from '../queuebox/queue-box-types.ts';
 import { EntityStatus, Key, NEVER_EXPIRE_TS, ResourceEntry } from '../queuebox/ResourceEntry.ts';
+import { computeResourceEntryFromALMessage } from './queue-box-resource-entry-computation.ts';
 
 export class QueueBoxUtilities {
     static readonly RETRY_DISPOSITION_ERROR = 'Queue entry requested retry';
